# spec/openapi.yaml
openapi: 3.1.0
info:
  title: OptraSight Express API
  version: 2.2.0
  summary: Minimal accurate contract for the active Express 5 + SQLite API.
  description: |
    OptraSight is served by one Express process on `/api/v1`. Authenticated
    routes use an opaque bearer session from `/auth/login`; MSSP admins may
    pivot tenant context with `X-Tenant-Id`. This file intentionally documents
    the currently supported public interface rather than old FastAPI/Celery
    claims.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Express server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Issue an opaque bearer session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
  schemas:
    AuthResponse:
      type: object
      required: [access_token, token_type, tenant_id, role, email]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: bearer }
        tenant_id: { type: string }
        role: { type: string }
        email: { type: string, format: email }